pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["EvilMadSquirrel <minichev.s.l@gmail.com>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.11b1", allow-prereleases = true}
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff_script:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
